examples(1_Advanced): + hiprtc interop example
In this example, we link a device code file that contains
a HIP C++ device kernel "__global__ void scale(float[],float)" that has an unresolved call
to a function "__device__ void scale_op(float[],float)" in its body with a
LLVM IR file that contains the definition of device function "scale_op".

To make this work, the HIP C++ snippet needs to be compiled with
the ``-fgpu-rdc`` option and the compilation results needs to
be added as `HIPRTC_JIT_INPUT_LLVM_BITCODE` type input to the link object.